Weaken inlining and match-substitution
The match-substitution transformation
  match e with
    C x y ->
      ... e ...
=>
  match e with
    C x y ->
      ... C x y ...
now only applies if e is a variable.

This causes problems when a large expression against twice,
e.g. as happens in the sorting examples:
  match (sort xs) with
    ... -> ... match (sort xs) ...
This can sometimes happen as a result of inlining. To stop this, the
inliner now only inlines values that look like they might lead to more
simplifications. This includes builtins, constructors and lambdas, but
not arbitrary function calls.