Relax monomorphisation requirements
Before, only this one would be created:

   (append n a) => (length n a) => (plus n) => (Decl_Assert_Lemma 0 n a),

Now this one is too:

   (append n a) => (length n a) => (Decl_Assert_Lemma 0 n a),

The reason is that plus has no type variables.
Further, the first rule is added in case the lemma depends on no
functions with type variables.
Todo: if only type-variable free functions imply a lemma, let any of them do it.